refactor/#22: 폴더 구조 변경 및 패키지 이동
