+++ app/src/main/java/org/sopt/starbanking/presentation/allAcounts/components/AccountCard.kt
@@ -1,4 +1,4 @@
-package org.sopt.starbanking.presentation.home.components
+package org.sopt.starbanking.presentation.allAcounts.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
