+++ app/src/main/java/org/sopt/starbanking/presentation/allAcounts/components/HomeTabMenu.kt
@@ -1,4 +1,4 @@
-package org.sopt.starbanking.presentation.home.components
+package org.sopt.starbanking.presentation.allAcounts.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
