--- app/src/main/java/org/sopt/starbanking/presentation/home/HomeViewModel.kt
@@ -1,2 +1,0 @@
-package org.sopt.starbanking.presentation.home
-
